+++ frontend/src/App.tsx
@@ -9,6 +9,7 @@ import DangKyTrungTam from '@/pages/DangKyTrungTam'
 import TongQuan from '@/pages/TongQuan'
 import LopHoc from '@/pages/dao-tao/LopHoc'
 import TaiLieu from '@/pages/dao-tao/TaiLieu'
+import HocPhi from '@/pages/dao-tao/HocPhi'
 import TaiKhoan from '@/pages/quan-tri/TaiKhoan'
 import PhanQuyen from '@/pages/quan-tri/PhanQuyen'
 import ThietLap from '@/pages/quan-tri/ThietLap'
@@ -58,6 +59,7 @@ export default function App() {
               <Route path="/" element={<TongQuan />} />
               <Route path="/lop-hoc" element={<LopHoc />} />
               <Route path="/tai-lieu" element={<TaiLieu />} />
+              <Route path="/hoc-phi" element={<HocPhi />} />
               <Route path="/quan-tri/tai-khoan" element={<TaiKhoan />} />
               <Route path="/quan-tri/phan-quyen" element={<PhanQuyen />} />
               <Route path="/quan-tri/thiet-lap" element={<ThietLap />} />